move add code to add mod
